+++ webknossos/Changelog.md
@@ -20,6 +20,8 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 
 ### Fixed
 
+- Fixed a bug in reading project info from webknossos using the api client for non-admins. [#972](https://github.com/scalableminds/webknossos-libs/pull/972)
+
 
 ## [0.14.12](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.14.12) - 2023-12-19
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.14.11...v0.14.12)